Refactoring to be able to use template argument deduction for destructor calls of members of structs
